release 3.0.20-beta source code for java

+++ services/ces/src/main/java/com/huaweicloud/sdk/ces/v1/model/AlarmTemplateCondition.java
@@ -52,7 +52,7 @@ public class AlarmTemplateCondition  {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value="value")
     
-    private Integer value;
+    private Double value;
 
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -162,7 +162,7 @@ public class AlarmTemplateCondition  {
         this.unit = unit;
     }
 
-    public AlarmTemplateCondition withValue(Integer value) {
+    public AlarmTemplateCondition withValue(Double value) {
         this.value = value;
         return this;
     }
@@ -174,11 +174,11 @@ public class AlarmTemplateCondition  {
      * 告警阈值，取值范围[0, Number.MAX_VALUE]，Number.MAX_VALUE值为1.7976931348623157e+108。具体阈值取值请参见附录中各服务监控指标中取值范围，如支持监控的服务列表中ECS的CPU使用率cpu_util取值范围可配置80。
      * @return value
      */
-    public Integer getValue() {
+    public Double getValue() {
         return value;
     }
 
-    public void setValue(Integer value) {
+    public void setValue(Double value) {
         this.value = value;
     }
